Rework how desktop sessions are shown on job page
Sessions are not part of the resutls and outputs.  The results and
outputs are items that are expected to only be of interest once the job
has completed, and to exist after the job is completed.  Sessions are
the opposite of that.

There is now a new "Desktop session" card below the metadata card.  It
is stripped down version of the session card in the desktop app.