Remove support for long polling
It isn't being used at the moment and having the (by default) single
thread occupied by the long poll is not a good idea.

We will eventually, want to inform the client more quickly when the
session is available, but that is likely best done with server sent
events or websockets in some fashion.