Workaround bad configuration breaking things
If an exception is raised whilst loading `config/puma.rb` the `pumactl`
script can become broken.  In particular when attempting to stop the
service.

However, we want bad configurations to prevent the service from
starting.  The code checking for bad configuration has been moved from
`config/boot.rb` to `config.ru` such that it will only run when the
service is started.