Fix two issues with select questions
If two select (multiselect) questions were asked in a row, the answer to
the first would overwrite the default to the second.  This was
particularly bad for multiselect where the answer might not even be
valid.

Fix bug with multiselect questions not having a default defined.
Previously a multiselect with a default of `null` would cause an error
preventing the script creation from continuing.