Index routes preload included resources
The JSON format used between job CLI and API does not yet have a
sensible mechanism to serialize included resources.  Neither does it
have a mechanism to query a subset of attributes.

This was resulting in some very large JSON payloads with duplicated
unneeded data.  We avoid duplicating the data by preloading the related
resources.